+++ stk_wrapper/stk/include/Delay.h
@@ -37,7 +37,7 @@ public:
   ~Delay();
 
   //! Get the maximum delay-line length.
-  unsigned long getMaximumDelay( void ) { return inputs_.size() - 1; };
+  unsigned long getMaximumDelay( void ) { return static_cast<unsigned long>(inputs_.size() - 1); };
 
   //! Set the maximum delay-line length.
   /*!
